add: deletion control
This feature is useful when used along with JWT tokens and X-Headers.
Instead of deletion a captcha entry after its solving, you can leave it in the storage and issue a token that might be used as a request header field and prove that you're allowed to procced that particular request. For example, this might be useful to create interactive captcha input interface: instead of sending captcha along with the rest data and getting 4xx code for entire request (in case if captcha was solved wrong), you can correctly solve captcha first and use obtained `X-Captcha` token for authorization the request.
An important thing to mention is that you have to care manually about deletion that captcha from the storage, otherwise it can be used multiple times before JWT token expires.